LayoutTests/fast/forms/validation-custom-message-expected.txt

Tests that the custom validation message set by JavaScript is displayed on UI side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS validationBubbleContents.message is "This is a custom validity message."
PASS validationBubbleContents.message is "Please fill out this field."
PASS successfullyParsed is true

TEST COMPLETE
Required text input: 
